requirements.txt
slack-bolt==1.18.0
flask==2.3.3
google-cloud-firestore==2.11.1
python-dotenv==1.0.0
openai==1.3.0
pytz==2023.3
schedule==1.2.0
urllib3<2.0.0  # To fix the OpenSSL warning
requests>=2.31.0
python-dateutil>=2.8.2 